requirements.txt
pymongo>=4.3.0
pyyaml>=6.0
docker>=6.0.0
aiohttp>=3.8.0
transformers>=4.25.0
torch>=2.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
psutil>=5.9.0
botocore>=1.29.0
daytona-sdk